+++ internal/adapters/https/pkg/app/init_loggers.go
@@ -1,8 +1,8 @@
 package app
 
-import hla_tcp_settings "github.com/number571/hidden-lake/internal/adapters/http/pkg/settings"
+import hla_https_settings "github.com/number571/hidden-lake/internal/adapters/https/pkg/settings"
 
 func (p *sApp) initLoggers() {
-	p.fIntAdapter.WithLogger(hla_tcp_settings.GetAppShortNameFMT(), p.fAnonLogger)
-	p.fExtAdapter.WithLogger(hla_tcp_settings.GetAppShortNameFMT(), p.fAnonLogger)
+	p.fIntAdapter.WithLogger(hla_https_settings.GetAppShortNameFMT()+"(INT)", p.fAnonLogger)
+	p.fExtAdapter.WithLogger(hla_https_settings.GetAppShortNameFMT(), p.fAnonLogger)
 }